Move things to .hpp
